don't unnecessarily copy blocks in DZCDataController
I thought a while ago that copying these blocks fixed crashes
when calling a block from inside itself. It seems like
adding __block to the block definition is the correct way to
do this, and [ copy]ing them was causing a memory leak.

I will test this for a while to see whether it introduces
any crashes.